+++ lib/easy-rest/caching/cache-item-removed-reason.d.ts
@@ -1,0 +1,5 @@
+export declare enum CacheItemRemovedReason {
+    expired = 0,
+    removed = 1,
+    underused = 2,
+}