+++ lib/index.d.ts
@@ -1,4 +1,5 @@
 export * from './easy-rest/api';
+export * from './easy-rest/caching';
 export * from './easy-rest/core';
 export * from './easy-rest/decorators';
 export * from './easy-rest/handlers';