+++ src/easy-rest/http/http-context.ts
@@ -1,8 +1,10 @@
 import {Request, Response, NextFunction} from "express";
 import {IPrincipal} from "../security/principal/principal";
+import {Cache} from "../caching/cache";
 
 export class HttpContext {
   user: IPrincipal;
+  cache: Cache;
 
   constructor(private req: Request, private res: Response, private nxt: NextFunction) {
   }